requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
drf-jwt==1.17.2
Markdown==3.3.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
